refactor(store): 统一数据库连接并重构逻辑数据库索引管理
将读写分离的连接池合并为单一连接，简化数据库层配置。引入上下文传递逻辑数据库索引，替代原有的 WithDB 方法，消除并发数据竞争。移除不再使用的辅助函数，并调整列表位置字段为 int64 以避免浮点数精度问题。同时限制事务队列大小防止内存耗尽，并优化后台清理协程的停止机制。

+++ internal/store/model.go
@@ -45,11 +45,12 @@ func (RHash) TableName() string { return "rhash" }
 
 // RList stores list elements with position.
 // Ordered by position within each key.
+// Position is an int64 to avoid float64 precision issues with large lists.
 type RList struct {
-	ID   int     `gorm:"column:id;primaryKey;autoIncrement"` // For Postgres/MySQL
-	KID  int     `gorm:"column:kid;uniqueIndex:rlist_uniq_idx,priority:1;not null"`
-	Pos  float64 `gorm:"column:pos;uniqueIndex:rlist_uniq_idx,priority:2;not null"`
-	Elem []byte  `gorm:"column:elem;not null"`
+	ID   int    `gorm:"column:id;primaryKey;autoIncrement"` // For Postgres/MySQL
+	KID  int    `gorm:"column:kid;uniqueIndex:rlist_uniq_idx,priority:1;not null"`
+	Pos  int64  `gorm:"column:pos;uniqueIndex:rlist_uniq_idx,priority:2;not null"`
+	Elem []byte `gorm:"column:elem;not null"`
 }
 
 // TableName returns the table name for GORM.
